renaming ShowIdentity
