+++ src/components/display/fewlines/ConfirmationBox/ConfirmationBox.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Button, ButtonVariant } from "../Button/Button";
 import { ClickAwayListener } from "../ClickAwayListener";
 import { Container } from "../Container";
-import { ConfirmationText } from "../ShowIdentity/ShowIdentity";
+import { ConfirmationText } from "../IdentityOverview/IdentityOverview";
 import { ConfirmationBox } from "./ConfirmationBox";
 
 export default {
